feat: Introduce AppError for improved error handling in handlers 🎉
- Added AppError enum to manage different error types.
- Updated client, invoice, session, and user profile handlers to use AppError for error responses.
- Refactored error handling in various service functions to return AppError.
- Changed default_hourly_rate type from f64 to f32 in client and invoice models for consistency.
- Removed unused schema file.

+++ backend/src/models/client.rs
@@ -9,7 +9,7 @@ pub struct Client {
     pub name: String,
     pub address: String,
     pub contact_person: Option<String>,
-    pub default_hourly_rate: f64,
+    pub default_hourly_rate: f32,
 }
 
 #[derive(Debug, Deserialize, Insertable)]
@@ -18,7 +18,7 @@ pub struct NewClient {
     pub name: String,
     pub address: String,
     pub contact_person: Option<String>,
-    pub default_hourly_rate: f64,
+    pub default_hourly_rate: f32,
 }
 
 #[derive(Debug, Deserialize, AsChangeset)]
@@ -27,5 +27,5 @@ pub struct UpdateClient {
     pub name: Option<String>,
     pub address: Option<String>,
     pub contact_person: Option<String>,
-    pub default_hourly_rate: Option<f64>,
+    pub default_hourly_rate: Option<f32>,
 }
